FEAT.회원 서비스 추가 및 검색 테스트 코드 추가
