+++ src/main/java/jpabook/jpashop/HelloController.java
@@ -1,4 +1,4 @@
-package japbook.jpashop;
+package jpabook.jpashop;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
