+++ src/main/java/jpabook/jpashop/domain/Address.java
@@ -1,4 +1,4 @@
-package japbook.jpashop.domain;
+package jpabook.jpashop.domain;
 
 import jakarta.persistence.Embeddable;
 import lombok.Getter;
