+++ src/main/java/jpabook/jpashop/domain/Item/Album.java
@@ -1,4 +1,4 @@
-package japbook.jpashop.domain.Item;
+package jpabook.jpashop.domain.Item;
 
 import jakarta.persistence.DiscriminatorValue;
 import jakarta.persistence.Entity;
@@ -6,11 +6,11 @@ import lombok.Getter;
 import lombok.Setter;
 
 @Entity
-@DiscriminatorValue("M")
+@DiscriminatorValue("A")
 @Getter @Setter
-public class Movie extends Item {
+public class Album extends Item {
 
-    private String director;
-    private String actor;
+    private String artist;
+    private String etc;
 
 }
