+++ src/main/java/jpabook/jpashop/domain/Item/Book.java
@@ -1,4 +1,4 @@
-package japbook.jpashop.domain.Item;
+package jpabook.jpashop.domain.Item;
 
 import jakarta.persistence.DiscriminatorValue;
 import jakarta.persistence.Entity;
@@ -6,11 +6,11 @@ import lombok.Getter;
 import lombok.Setter;
 
 @Entity
-@DiscriminatorValue("A")
+@DiscriminatorValue("B")
 @Getter @Setter
-public class Album extends Item {
+public class Book extends Item {
 
-    private String artist;
-    private String etc;
+    private String author;
+    private String isbn;
 
 }
