+++ src/main/java/jpabook/jpashop/domain/Item/Movie.java
@@ -1,4 +1,4 @@
-package japbook.jpashop.domain.Item;
+package jpabook.jpashop.domain.Item;
 
 import jakarta.persistence.DiscriminatorValue;
 import jakarta.persistence.Entity;
@@ -6,11 +6,11 @@ import lombok.Getter;
 import lombok.Setter;
 
 @Entity
-@DiscriminatorValue("B")
+@DiscriminatorValue("M")
 @Getter @Setter
-public class Book extends Item {
+public class Movie extends Item {
 
-    private String author;
-    private String isbn;
+    private String director;
+    private String actor;
 
 }
