+++ src/main/java/jpabook/jpashop/domain/Member.java
@@ -1,4 +1,4 @@
-package japbook.jpashop.domain;
+package jpabook.jpashop.domain;
 
 import java.util.ArrayList;
 import java.util.List;
