+++ src/main/java/jpabook/jpashop/domain/service/MemberService.java
@@ -1,9 +1,8 @@
-package japbook.jpashop.domain.service;
+package jpabook.jpashop.domain.service;
 
-import japbook.jpashop.domain.Member;
-import japbook.jpashop.repository.MemberRepository;
+import jpabook.jpashop.domain.Member;
+import jpabook.jpashop.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -90,8 +89,8 @@ public class MemberService {
      */
     private void validationDuplicateMember(Member member) {
         List<Member> findMembers = memberRepository.findByName(member.getName());
-        // if(!findMembers.isEmpty())
-        if (findMembers.size() > 1) {
+        if(!findMembers.isEmpty()) {
+        //if (findMembers.size() > 1) {
             throw new IllegalStateException("이미 존재하는 회원입니다.");
         }
     }
