+++ src/test/java/jpabook/jpashop/JpashopApplicationTests.java
@@ -1,4 +1,4 @@
-package japbook.jpashop;
+package jpabook.jpashop;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
